+++ app/Http/Controllers/Customer/CustomerController.php
@@ -2,20 +2,14 @@
 
 namespace App\Http\Controllers\Customer;
 
-use App\DTO\User\CustomerDTO;
 use App\Helpers\Helpers;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use App\Models\User\User;
-
 use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
-use App\Http\Requests\CustomerRequests\UpdateCustomerAddressRequest;
 use App\Http\Requests\CustomerRequests\AddFavoriteRestaurantRequest;
 use App\Http\Requests\CustomerRequests\SubmitFeedbackRequest;
-use App\Http\Requests\CustomerRequests\UpdateProfileRequest;
 use App\Http\Requests\CustomerRequests\UsePointsRequest;
-use App\Http\Resources\Customer\ProfileResource;
 use App\Services\Customer\CustomerService;
 
 class CustomerController extends Controller
@@ -30,17 +24,6 @@ class CustomerController extends Controller
             $user = Auth::user();
             $this->customer = $user->customer;
         }
-    }
-
-    public function editProfile(UpdateProfileRequest $request)
-    {
-        $userId = auth()->user()->id;
-
-        $this->customerService->updateProfile($userId, $request);
-
-        $updatedUser = User::with('customer')->find($userId);
-
-        return Helpers::sendSuccessResponse(Response::HTTP_OK, 'Profile updated successfully', $updatedUser);
     }
 
     public function viewMenus($restaurantId)
@@ -77,43 +60,6 @@ class CustomerController extends Controller
         return Helpers::sendSuccessResponse(Response::HTTP_OK, 'Points redeemed successfully', ['monetary_value' => $monetaryValue]);
     }
 
-
-    public function updateCustomerAddress(UpdateCustomerAddressRequest $request)
-    {
-        $customerId = $request->get('customer_id');
-        $isDefaultAddress = $request->routeIs('updateCustomerDefaultAddress');
-
-        $data = [
-            'user_id' => $customerId,
-            'address' => $validatedData['address'] ?? null,
-            'delivery_address' => $isDefaultAddress ? null : ($validatedData['delivery_address'] ?? null),
-            'favorites' => $validatedData['favorites'] ?? null
-        ];
-
-        // Ensure either address or delivery_address is present
-        if (is_null($data['address']) && is_null($data['delivery_address'])) {
-            return Helpers::sendFailureResponse(Response::HTTP_BAD_REQUEST, 'Either address or delivery address must be provided.');
-        }
-
-        $customerDTO = new CustomerDTO($data);
-
-        $this->customerService->updateCustomerInfo($customerId, $customerDTO);
-
-        return Helpers::sendSuccessResponse(Response::HTTP_OK, 'Customer address updated successfully');
-    }
-
-    public function viewProfile()
-    {
-        $userId = auth()->user()->id;
-        $customer = $this->customerService->getProfile($userId);
-
-        return Helpers::sendSuccessResponse(
-            Response::HTTP_OK,
-            'Customer profile retrieved successfully',
-            new ProfileResource($customer)
-        );
-    }
-
     public function addFavoriteRestaurant(AddFavoriteRestaurantRequest $request)
     {
         $restaurantId = $request->get('restaurant_id');
@@ -146,11 +92,6 @@ class CustomerController extends Controller
 
         return Helpers::sendSuccessResponse(Response::HTTP_OK, 'All restaurants retrieved successfully', $restaurants);
     }
-    // public function viewRestaurantById(Request $request)
-    // {
-    //     $restaurant = $this->customerService->viewRestaurantById($request->all());
-    //     return Helpers::sendSuccessResponse(Response::HTTP_OK, 'Restaurant retrieved successfully', $restaurant);
-    // }
 
     public function viewDeals()
     {
