+++ app/Http/Resources/Customer/OrderDetailsResource.php
@@ -11,7 +11,7 @@ class OrderDetailsResource extends JsonResource
     {
         return [
             'order_id' => $this->id,
-            'order_image' => $this->restaurant->logo_path ? Storage::url($this->restaurant->logo_path) : null,
+            'order_image' => rtrim(env('APP_URL'), '/') . '/' . ltrim(Storage::url($this->restaurant->logo_path), '/'),
             'restaurant_name' => $this->restaurant->name ?? 'N/A',
             'restaurant_branch_address' => $this->branch ? $this->branch->address : 'N/A',
             'customer_address' => $this->customer ? $this->customer->delivery_address ?? 'N/A' : 'N/A',
